Reset `base::Thread::delegate_` on main thread only
This commit resets the `reset()` of `base::Thread::delegate_` to be on
the main thread only, instead of on the backing thread when it
terminates. The previous design was bad because it meant the API
surface of `base::Thread` (`GetTaskRunner()` specifically) could
change unpredictably by the backing thread terminating itself. This new
design ensures that `base::Thread::delegate_` is only ever mutated on
the main thread, and clarifies that this is thread-safe because it is
only ever reset after the backing thread is terminated/joined.